pyproject.toml
[project]
name = "slack-mcp-client"
version = "0.1.0"
description = "A client for the Model Context Protocol (MCP) that can be used with Slack and multiple AI providers"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.51.0",
    "python-dotenv>=1.0.0",
    "mcp>=1.9.0",
    "openai>=1.0.0",
    "google-generativeai>=0.3.0",
    "google-genai>=1.16.1",
    "slack-bolt>=1.23.0",
    "pyyaml>=6.0.2",
    "aiohttp>=3.12.2",
]


[project.scripts]
mcp-client = "mcp_client.cli:run"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
